requirements.txt
# pip install -r requirements.txt
wheel
sqlalchemy>=1.3.19
pytest>=6.1.0
setuptools>=49.6.0
click>=7.1.2
flask>=1.1.2
xlrd>=1.2.0
pandas>=1.1.3
openpyxl
# finnhub-python
flask-sqlalchemy>=2.4.4
flask-cors>=3.0.9
Werkzeug>=1.0.1
numpy>=1.16.2
pyyaml
flask-restx
pyjwt>=1.7.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
XlsxWriter
